integrating search
