+++ frontend/app/src/i18n/en.json
@@ -879,7 +879,6 @@
         "invite": "Invite",
         "imageLabel": "Banner/Avatar image",
         "name": "Name",
-        "nameRules": "Community names must be unique",
         "namePlaceholder": "Enter community name",
         "description": "Description",
         "descriptionPlaceholder": "Tell us what your community is all about",
