+++ frontend/openchat-agent/src/services/group/candid/idl.js
@@ -1103,7 +1103,7 @@ export const idlFactory = ({ IDL }) => {
     'SetToNone' : IDL.Null,
     'SetToSome' : Milliseconds,
   });
-  const AvatarIdUpdate = IDL.Variant({
+  const DocumentIdUpdate = IDL.Variant({
     'NoChange' : IDL.Null,
     'SetToNone' : IDL.Null,
     'SetToSome' : IDL.Nat,
@@ -1132,7 +1132,7 @@ export const idlFactory = ({ IDL }) => {
     'description' : IDL.Opt(IDL.Text),
     'events_ttl' : EventsTimeToLiveUpdate,
     'last_updated' : TimestampMillis,
-    'avatar_id' : AvatarIdUpdate,
+    'avatar_id' : DocumentIdUpdate,
     'next_message_expiry' : TimestampUpdate,
     'latest_threads' : IDL.Vec(GroupCanisterThreadDetails),
     'frozen' : FrozenGroupUpdate,
@@ -1228,15 +1228,15 @@ export const idlFactory = ({ IDL }) => {
     'reply_in_thread' : IDL.Opt(PermissionRole),
     'react_to_messages' : IDL.Opt(PermissionRole),
   });
-  const Avatar = IDL.Record({
+  const Document = IDL.Record({
     'id' : IDL.Nat,
     'data' : IDL.Vec(IDL.Nat8),
     'mime_type' : IDL.Text,
   });
-  const AvatarUpdate = IDL.Variant({
+  const DocumentUpdate = IDL.Variant({
     'NoChange' : IDL.Null,
     'SetToNone' : IDL.Null,
-    'SetToSome' : Avatar,
+    'SetToSome' : Document,
   });
   const UpdateGroupV2Args = IDL.Record({
     'permissions' : IDL.Opt(OptionalGroupPermissions),
@@ -1246,7 +1246,7 @@ export const idlFactory = ({ IDL }) => {
     'events_ttl' : EventsTimeToLiveUpdate,
     'correlation_id' : IDL.Nat64,
     'rules' : IDL.Opt(AccessRules),
-    'avatar' : AvatarUpdate,
+    'avatar' : DocumentUpdate,
   });
   const FieldTooLongResult = IDL.Record({
     'length_provided' : IDL.Nat32,
