+++ frontend/openchat-agent/src/services/groupIndex/candid/idl.d.ts
@@ -6,11 +6,13 @@ import {
     FreezeGroupResponse,
     GroupMatch,
     RecommendedGroupsResponse,
-    SearchResponse,
+    SearchV2Response,
     UnfreezeGroupResponse,
     AddHotGroupExclusionResponse,
     RemoveHotGroupExclusionResponse,
-    SetUpgradeConcurrencyResponse
+    SetUpgradeConcurrencyResponse,
+    CommunityMatch,
+    SearchScope,
 } from "./types";
 export {
     _SERVICE as GroupIndexService,
@@ -19,11 +21,13 @@ export {
     FreezeGroupResponse as ApiFreezeGroupResponse,
     GroupMatch as ApiGroupMatch,
     RecommendedGroupsResponse as ApiRecommendedGroupsResponse,
-    SearchResponse as ApiSearchResponse,
+    SearchV2Response as ApiSearchResponse,
     UnfreezeGroupResponse as ApiUnfreezeGroupResponse,
     AddHotGroupExclusionResponse as ApiAddHotGroupExclusionResponse,
     RemoveHotGroupExclusionResponse as ApiRemoveHotGroupExclusionResponse,
     SetUpgradeConcurrencyResponse as ApiSetUpgradeConcurrencyResponse,
+    CommunityMatch as ApiCommunityMatch,
+    SearchScope as ApiSearchScope,
 };
 
 export const idlFactory: IDL.InterfaceFactory;
