+++ frontend/openchat-agent/src/services/groupIndex/candid/idl.js
@@ -1,6 +1,38 @@
 export const idlFactory = ({ IDL }) => {
   const CanisterId = IDL.Principal;
+  const CommunityId = CanisterId;
+  const TimestampMillis = IDL.Nat64;
   const ChatId = CanisterId;
+  const ActiveGroupsArgs = IDL.Record({
+    'community_ids' : IDL.Vec(CommunityId),
+    'active_since' : IDL.Opt(TimestampMillis),
+    'group_ids' : IDL.Vec(ChatId),
+  });
+  const UserId = CanisterId;
+  const DeletedCommunityInfo = IDL.Record({
+    'id' : CommunityId,
+    'name' : IDL.Text,
+    'public' : IDL.Bool,
+    'timestamp' : TimestampMillis,
+    'deleted_by' : UserId,
+  });
+  const DeletedGroupInfo = IDL.Record({
+    'id' : ChatId,
+    'name' : IDL.Text,
+    'public' : IDL.Bool,
+    'timestamp' : TimestampMillis,
+    'deleted_by' : UserId,
+    'group_name' : IDL.Text,
+  });
+  const ActiveGroupsResponse = IDL.Variant({
+    'Success' : IDL.Record({
+      'deleted_communities' : IDL.Vec(DeletedCommunityInfo),
+      'deleted_groups' : IDL.Vec(DeletedGroupInfo),
+      'active_groups' : IDL.Vec(ChatId),
+      'timestamp' : TimestampMillis,
+      'active_communities' : IDL.Vec(CommunityId),
+    }),
+  });
   const AddHotGroupExclusionArgs = IDL.Record({ 'chat_id' : ChatId });
   const AddHotGroupExclusionResponse = IDL.Variant({
     'ChatAlreadyExcluded' : IDL.Null,
@@ -17,18 +49,9 @@ export const idlFactory = ({ IDL }) => {
     'ChatNotFrozen' : IDL.Null,
     'InternalError' : IDL.Text,
   });
-  const TimestampMillis = IDL.Nat64;
   const FilterGroupsArgs = IDL.Record({
     'active_since' : IDL.Opt(TimestampMillis),
     'chat_ids' : IDL.Vec(ChatId),
-  });
-  const UserId = CanisterId;
-  const DeletedGroupInfo = IDL.Record({
-    'id' : ChatId,
-    'public' : IDL.Bool,
-    'timestamp' : TimestampMillis,
-    'deleted_by' : UserId,
-    'group_name' : IDL.Text,
   });
   const FilterGroupsResponse = IDL.Variant({
     'Success' : IDL.Record({
@@ -461,6 +484,33 @@ export const idlFactory = ({ IDL }) => {
     'TermTooLong' : IDL.Nat8,
     'InvalidTerm' : IDL.Null,
   });
+  const SearchScope = IDL.Variant({
+    'All' : IDL.Null,
+    'Communities' : IDL.Null,
+    'Groups' : IDL.Null,
+  });
+  const SearchV2Args = IDL.Record({
+    'max_results' : IDL.Nat8,
+    'scope' : SearchScope,
+    'search_term' : IDL.Text,
+  });
+  const CommunityMatch = IDL.Record({
+    'id' : CommunityId,
+    'name' : IDL.Text,
+    'description' : IDL.Text,
+    'avatar_id' : IDL.Opt(IDL.Nat),
+    'banner_id' : IDL.Opt(IDL.Nat),
+  });
+  const SearchV2Success = IDL.Record({
+    'group_matches' : IDL.Vec(GroupMatch),
+    'community_matches' : IDL.Vec(CommunityMatch),
+  });
+  const SearchV2Response = IDL.Variant({
+    'TermTooShort' : IDL.Nat8,
+    'Success' : SearchV2Success,
+    'TermTooLong' : IDL.Nat8,
+    'InvalidTerm' : IDL.Null,
+  });
   const SetUpgradeConcurrencyArgs = IDL.Record({ 'value' : IDL.Nat32 });
   const SetUpgradeConcurrencyResponse = IDL.Variant({
     'NotAuthorized' : IDL.Null,
@@ -482,6 +532,11 @@ export const idlFactory = ({ IDL }) => {
     'InternalError' : IDL.Text,
   });
   return IDL.Service({
+    'active_groups' : IDL.Func(
+        [ActiveGroupsArgs],
+        [ActiveGroupsResponse],
+        ['query'],
+      ),
     'add_hot_group_exclusion' : IDL.Func(
         [AddHotGroupExclusionArgs],
         [AddHotGroupExclusionResponse],
@@ -509,6 +564,7 @@ export const idlFactory = ({ IDL }) => {
         [],
       ),
     'search' : IDL.Func([SearchArgs], [SearchResponse], ['query']),
+    'search_v2' : IDL.Func([SearchV2Args], [SearchV2Response], ['query']),
     'set_community_upgrade_concurrency' : IDL.Func(
         [SetUpgradeConcurrencyArgs],
         [SetUpgradeConcurrencyResponse],
