+++ frontend/openchat-agent/src/services/user/candid/types.d.ts
@@ -44,22 +44,16 @@ export interface AudioContent {
   'blob_reference' : [] | [BlobReference],
   'caption' : [] | [string],
 }
-export interface Avatar {
-  'id' : bigint,
-  'data' : Uint8Array | number[],
-  'mime_type' : string,
-}
 export interface AvatarChanged {
   'changed_by' : UserId,
   'previous_avatar' : [] | [bigint],
   'new_avatar' : [] | [bigint],
 }
-export type AvatarIdUpdate = { 'NoChange' : null } |
-  { 'SetToNone' : null } |
-  { 'SetToSome' : bigint };
-export type AvatarUpdate = { 'NoChange' : null } |
-  { 'SetToNone' : null } |
-  { 'SetToSome' : Avatar };
+export interface BannerChanged {
+  'new_banner' : [] | [bigint],
+  'changed_by' : UserId,
+  'previous_banner' : [] | [bigint],
+}
 export type BioArgs = {};
 export type BioResponse = { 'Success' : string };
 export interface BlobReference {
@@ -188,6 +182,7 @@ export interface CommunityCanisterChannelSummary {
   'next_message_expiry' : [] | [TimestampMillis],
   'latest_threads' : Array<GroupCanisterThreadDetails>,
   'latest_event_index' : EventIndex,
+  'banner_id' : [] | [bigint],
   'history_visible_to_new_joiners' : boolean,
   'min_visible_message_index' : MessageIndex,
   'mentions' : Array<Mention>,
@@ -210,7 +205,7 @@ export interface CommunityCanisterChannelSummaryUpdates {
   'description' : [] | [string],
   'events_ttl' : EventsTimeToLiveUpdate,
   'last_updated' : TimestampMillis,
-  'avatar_id' : AvatarIdUpdate,
+  'avatar_id' : DocumentIdUpdate,
   'latest_threads' : Array<GroupCanisterThreadDetails>,
   'latest_event_index' : [] | [EventIndex],
   'mentions' : Array<Mention>,
@@ -232,6 +227,7 @@ export interface CommunityCanisterCommunitySummary {
   'avatar_id' : [] | [bigint],
   'frozen' : [] | [FrozenGroupInfo],
   'latest_event_index' : EventIndex,
+  'banner_id' : [] | [bigint],
   'member_count' : number,
 }
 export interface CommunityCanisterCommunitySummaryUpdates {
@@ -245,13 +241,14 @@ export interface CommunityCanisterCommunitySummaryUpdates {
   'description' : [] | [string],
   'last_updated' : TimestampMillis,
   'channels_removed' : Array<ChannelId>,
-  'avatar_id' : AvatarIdUpdate,
+  'avatar_id' : DocumentIdUpdate,
   'channels_added' : Array<CommunityCanisterChannelSummary>,
   'frozen' : FrozenGroupUpdate,
   'latest_event_index' : [] | [EventIndex],
+  'banner_id' : DocumentIdUpdate,
   'member_count' : [] | [number],
 }
-export type CommunityId = Principal;
+export type CommunityId = CanisterId;
 export interface CommunityMessagesRead {
   'community_id' : CommunityId,
   'channels_read' : Array<ChannelMessagesRead>,
@@ -284,11 +281,12 @@ export interface CreateCommunityArgs {
   'permissions' : [] | [CommunityPermissions],
   'gate' : [] | [AccessGate],
   'name' : string,
+  'banner' : [] | [Document],
   'description' : string,
   'history_visible_to_new_joiners' : boolean,
   'default_channels' : Array<string>,
   'rules' : AccessRules,
-  'avatar' : [] | [Avatar],
+  'avatar' : [] | [Document],
 }
 export type CreateCommunityResponse = { 'DefaultChannelsInvalid' : null } |
   { 'NameReserved' : null } |
@@ -304,7 +302,8 @@ export type CreateCommunityResponse = { 'DefaultChannelsInvalid' : null } |
   { 'NameTooLong' : FieldTooLongResult } |
   { 'NameTaken' : null } |
   { 'InternalError' : null } |
-  { 'MaxCommunitiesCreated' : number };
+  { 'MaxCommunitiesCreated' : number } |
+  { 'BannerTooBig' : FieldTooLongResult };
 export interface CreateCommunitySuccessResult { 'community_id' : CommunityId }
 export interface CreateGroupArgs {
   'is_public' : boolean,
@@ -314,7 +313,7 @@ export interface CreateGroupArgs {
   'description' : string,
   'history_visible_to_new_joiners' : boolean,
   'rules' : AccessRules,
-  'avatar' : [] | [Avatar],
+  'avatar' : [] | [Document],
 }
 export type CreateGroupResponse = { 'NameReserved' : null } |
   { 'RulesTooLong' : FieldTooLongResult } |
@@ -437,6 +436,17 @@ export interface DirectReactionAddedNotification {
   'timestamp' : TimestampMillis,
   'reaction' : string,
 }
+export interface Document {
+  'id' : bigint,
+  'data' : Uint8Array | number[],
+  'mime_type' : string,
+}
+export type DocumentIdUpdate = { 'NoChange' : null } |
+  { 'SetToNone' : null } |
+  { 'SetToSome' : bigint };
+export type DocumentUpdate = { 'NoChange' : null } |
+  { 'SetToNone' : null } |
+  { 'SetToSome' : Document };
 export type EditMessageResponse = { 'MessageNotFound' : null } |
   { 'ChatNotFound' : null } |
   { 'Success' : null } |
@@ -579,7 +589,7 @@ export interface GroupCanisterGroupChatSummaryUpdates {
   'description' : [] | [string],
   'events_ttl' : EventsTimeToLiveUpdate,
   'last_updated' : TimestampMillis,
-  'avatar_id' : AvatarIdUpdate,
+  'avatar_id' : DocumentIdUpdate,
   'next_message_expiry' : TimestampUpdate,
   'latest_threads' : Array<GroupCanisterThreadDetails>,
   'frozen' : FrozenGroupUpdate,
@@ -1309,7 +1319,7 @@ export type SendMessageWithTransferToGroupResponse = {
   { 'TransferFailed' : string } |
   { 'InternalError' : [string, CompletedCryptoTransaction] } |
   { 'CryptocurrencyNotSupported' : Cryptocurrency };
-export interface SetAvatarArgs { 'avatar' : [] | [Avatar] }
+export interface SetAvatarArgs { 'avatar' : [] | [Document] }
 export type SetAvatarResponse = { 'AvatarTooBig' : FieldTooLongResult } |
   { 'Success' : null } |
   { 'UserSuspended' : null };
@@ -1473,7 +1483,7 @@ export type UpdatesV2Response = {
       'direct_chats_added' : Array<DirectChatSummary>,
       'blocked_users_v2' : [] | [Array<UserId>],
       'group_chats_added' : Array<UserCanisterGroupChatSummary>,
-      'avatar_id' : AvatarIdUpdate,
+      'avatar_id' : DocumentIdUpdate,
       'chats_removed' : Array<ChatId>,
       'timestamp' : TimestampMillis,
       'group_chats_updated' : Array<UserCanisterGroupChatSummaryUpdates>,
