+++ frontend/openchat-client/src/openchat.ts
@@ -251,7 +251,7 @@ import {
     type CurrentUserResponse,
     type RemoveMemberResponse,
     type RegisterProposalVoteResponse,
-    type GroupSearchResponse,
+    type SearchResponse,
     type GroupInvite,
     type SearchDirectChatResponse,
     type SearchGroupChatResponse,
@@ -304,6 +304,7 @@ import {
     E8S_PER_TOKEN,
     Community,
     CreateCommunityResponse,
+    SearchScope,
 } from "openchat-shared";
 import { failedMessagesStore } from "./stores/failedMessages";
 import {
@@ -3221,8 +3222,12 @@ export class OpenChat extends OpenChatAgentWorker {
         return this.sendRequest({ kind: "getGroupRules", chatId });
     }
 
-    searchGroups(searchTerm: string, maxResults = 10): Promise<GroupSearchResponse> {
-        return this.sendRequest({ kind: "searchGroups", searchTerm, maxResults });
+    searchGroups(searchTerm: string, maxResults = 10): Promise<SearchResponse> {
+        return this.sendRequest({ kind: "search", searchTerm, maxResults, scope: "groups" });
+    }
+
+    searchCommunities(searchTerm: string, maxResults = 10): Promise<SearchResponse> {
+        return this.sendRequest({ kind: "search", searchTerm, maxResults, scope: "communities" });
     }
 
     dismissRecommendation(chatId: string): Promise<void> {
